Update solution for batch-normalization

+++ batch-normalization/batch-normalization.py
@@ -5,9 +5,9 @@ def batch_norm_forward(x, gamma, beta, eps=1e-5):
     Forward-only BatchNorm for (N,D) or (N,C,H,W).
     """
     # Write code here
-    x = np.asarray(x, float)
-    gamma = np.asarray(gamma, float)
-    beta = np.asarray(beta, float)
+    x = np.asarray(x, np.float64) 
+    gamma = np.asarray(gamma, np.float64)
+    beta = np.asarray(beta, np.float64)
 
     if x.ndim==2:
         mu = np.mean(x, keepdims=True, axis=0)
@@ -18,7 +18,6 @@ def batch_norm_forward(x, gamma, beta, eps=1e-5):
         var = np.var(x, axis=(0,2,3), keepdims=True)
         gamma = gamma.reshape(1,-1,1,1)
         beta = beta.reshape(1,-1,1,1)
-    
 
     x_hat = (x-mu) / np.sqrt(var + eps)
